# docs/openapi.yaml
openapi: 3.0.3
info:
  title: AxionPAY Payment API
  version: 0.2.0
servers:
  - url: http://localhost:3060
...
security:
  - ApiKeyAuth: []
    BearerAuth: []
paths:
...
  /payments/{id}:
    get:
      tags: [Payments]
      summary: Get payment by id
      parameters:
        - $ref: "#/components/parameters/PaymentId"
      responses:
        "200":
          description: Payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
...
components:
  parameters:
...
    PaymentId:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
    BearerAuth:
      type: http
      scheme: bearer
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      required: [ok, message]
      properties:
        ok: { type: boolean, example: false }
        error: { type: string }
        message: { type: string }
        details:
          type: object
          additionalProperties: true
...
    PaymentMethod:
      type: string
      enum: [pix, card]
    PaymentStatus:
      type: string
      enum: [pending, authorized, paid, failed, canceled, refunded, expired]
...
    Transaction:
      type: object
      properties:
        id: { type: string }
        amount: { type: number }
        amount_cents: { type: integer }
        currency: { type: string }
        method:
          $ref: "#/components/schemas/PaymentMethod"
        status:
          $ref: "#/components/schemas/PaymentStatus"
        capture: { type: boolean }
        customer:
          type: object
          nullable: true
        provider:
          type: string
          nullable: true
        providerReference:
          type: string
          nullable: true
        methodDetails:
          type: object
          nullable: true
        metadata:
          type: object
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        events:
          type: array
          items:
            $ref: "#/components/schemas/PaymentEvent"
    PaymentEvent:
      type: object
      properties:
        type: { type: string }
        at:
          type: string
          format: date-time
    PaymentResponse:
      type: object
      properties:
        ok: { type: boolean }
        transaction:
          $ref: "#/components/schemas/Transaction"
...